ai-backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
sse-starlette>=1.8.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
supabase>=2.0.0
python-dotenv>=1.0.0
slowapi>=0.1.9
